move firebase instance back to popup to get data persistency
