+++ component/client/chat/ClientChat.tsx
@@ -41,7 +41,7 @@ class ClientChat extends React.Component<IClientChatProps, IClientChatState> {
     console.log("[ INFO ] : ClientChat render url and messages", this.props.url, this.state.messages);
     return (
       <div>
-        <Panel header="Chat" bsStyle="primary" footer={this.messenger}>
+        <Panel header={"Chat at "+this.props.url} bsStyle="primary" footer={this.messenger}>
           <ChatHistory messages={this.state.messages} />
         </Panel>
       </div>
