+++ component/client/service/RoomService.tsx
@@ -1,6 +1,4 @@
 import * as firebase from "firebase";
-import {firebaseDB} from "../backgroundContext";
-
 import IData from "./IData";
 
 class RoomService {
@@ -11,7 +9,7 @@ class RoomService {
 
   constructor(url: string, onMessagePosted: (data: IData, from: any) => void) {
     const cleanUrl = url.replace(/[\\.]/g, ",");
-    this.rootRef = firebaseDB.ref(cleanUrl);
+    this.rootRef = firebase.database().ref(cleanUrl);
     this.messageRef = this.rootRef.child("message");
 
     // set up message reference
